cleanup files properly

+++ tests/unit/test_source_finder.py
@@ -157,7 +157,8 @@ def test_load_globals():
     sfinder.load_globals(filename, rms=1, mask=Region())
     if sfinder.region is None:
         raise AssertionError()
-    
+
+
 def test_load_globals_cube():
     """Test load_globals"""
     sfinder = sf.SourceFinder()
@@ -165,13 +166,14 @@ def test_load_globals_cube():
     # aux_files = sf.get_aux_files("tests/test_files/synthetic_with_alpha.fits")
     bkg = filename
     rms = filename
-    sfinder.load_globals(filename, bkgin=bkg, rmsin= rms, as_cube=True)
+    sfinder.load_globals(filename, bkgin=bkg, rmsin=rms, as_cube=True)
     if sfinder.img is None:
         raise AssertionError()
     if np.allclose(sfinder.img, 0):
         print("True")
     else:
         raise AssertionError()
+
 
 def test_find_and_prior_sources():
     """Test find sources and prior sources"""
@@ -263,7 +265,8 @@ def test_find_and_prior_sources():
         if not (os.path.exists("dlme")):
             raise AssertionError("Failed to create output file")
     finally:
-        os.remove("dlme")
+        if os.path.exists("dlme"):
+            os.remove("dlme")
 
 
 def dont_test_find_and_prior_parallel():
